+++ Assets/Features/Building/BuildableObjects/_Building_Atlas.asset
@@ -15,40 +15,58 @@ MonoBehaviour:
   Tiles:
   - TileType: 0
     TileData:
-      rid: 7365152698381369834
+      rid: 7365152724195475464
   - TileType: 0
     TileData:
-      rid: 7365152698381369839
+      rid: 7365152724195475465
   - TileType: 0
     TileData:
-      rid: 7365152698381369842
+      rid: 7365152724195475466
   - TileType: 1
     TileData:
-      rid: 7365152698381369836
+      rid: 7365152724195475467
   - TileType: 1
     TileData:
-      rid: 7365152698381369848
+      rid: 7365152724195475468
   - TileType: 1
     TileData:
-      rid: 7365152698381369850
+      rid: 7365152724195475469
   - TileType: 0
     TileData:
-      rid: 7365152698381369845
+      rid: 7365152724195475470
   - TileType: 0
     TileData:
-      rid: 7365152698381369852
+      rid: 7365152724195475471
   references:
     version: 2
     RefIds:
-    - rid: 7365152698381369834
-      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
+    - rid: 7365152724195475464
+      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
       data:
         Tile: {fileID: 11400000, guid: 29fa9a4e0e09c14478da003d3f391e1c, type: 2}
         Color: {r: 1, g: 1, b: 1, a: 1}
         BuildLoad: 2
         UnTraversable: 0
-    - rid: 7365152698381369836
-      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
+    - rid: 7365152724195475465
+      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
+      data:
+        Tile: {fileID: 11400000, guid: 6c780cbf86e291d499131d526e833735, type: 2}
+        Color: {r: 1, g: 1, b: 1, a: 1}
+        BuildLoad: 2
+        UnTraversable: 0
+    - rid: 7365152724195475466
+      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
+      data:
+        Tile: {fileID: 11400000, guid: 9ec80e97310e10147a0331cc257d09ed, type: 2}
+        Color: {r: 1, g: 1, b: 1, a: 1}
+        BuildLoad: 2
+        UnTraversable: 0
+    - rid: 7365152724195475467
+      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
       data:
         Tile: {fileID: 11400000, guid: 241963d8cfdb21c489ff7af6e2160525, type: 2}
         Color: {r: 1, g: 1, b: 1, a: 1}
@@ -56,29 +74,9 @@ MonoBehaviour:
         UnTraversable: 0
         UtilityType: 2
         Workload: 2
-    - rid: 7365152698381369839
-      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
-      data:
-        Tile: {fileID: 11400000, guid: fbfe88272625d5c47979b613d9233838, type: 2}
-        Color: {r: 1, g: 1, b: 1, a: 1}
-        BuildLoad: 2
-        UnTraversable: 1
-    - rid: 7365152698381369842
-      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
-      data:
-        Tile: {fileID: 11400000, guid: 188e4c7d9c6ce694a8392e304280d246, type: 2}
-        Color: {r: 1, g: 1, b: 1, a: 1}
-        BuildLoad: 2
-        UnTraversable: 0
-    - rid: 7365152698381369845
-      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
-      data:
-        Tile: {fileID: 11400000, guid: 6c780cbf86e291d499131d526e833735, type: 2}
-        Color: {r: 1, g: 1, b: 1, a: 1}
-        BuildLoad: 2
-        UnTraversable: 0
-    - rid: 7365152698381369848
-      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
+    - rid: 7365152724195475468
+      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
       data:
         Tile: {fileID: 11400000, guid: a34c6e45faeba8a4c9340342d4aa948a, type: 2}
         Color: {r: 1, g: 1, b: 1, a: 1}
@@ -86,8 +84,9 @@ MonoBehaviour:
         UnTraversable: 0
         UtilityType: 1
         Workload: 2
-    - rid: 7365152698381369850
-      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
+    - rid: 7365152724195475469
+      type: {class: UtilityTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
       data:
         Tile: {fileID: 11400000, guid: f31080f96096c074fb0adac6a5e51f4d, type: 2}
         Color: {r: 1, g: 1, b: 1, a: 1}
@@ -95,10 +94,19 @@ MonoBehaviour:
         UnTraversable: 0
         UtilityType: 3
         Workload: 2
-    - rid: 7365152698381369852
-      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes, asm: Assembly-CSharp}
+    - rid: 7365152724195475470
+      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
       data:
-        Tile: {fileID: 11400000, guid: 9ec80e97310e10147a0331cc257d09ed, type: 2}
+        Tile: {fileID: 11400000, guid: 188e4c7d9c6ce694a8392e304280d246, type: 2}
+        Color: {r: 1, g: 1, b: 1, a: 1}
+        BuildLoad: 2
+        UnTraversable: 0
+    - rid: 7365152724195475471
+      type: {class: NormalTile, ns: Features.Building.Scripts.Datatypes.TileData,
+        asm: Assembly-CSharp}
+      data:
+        Tile: {fileID: 11400000, guid: fbfe88272625d5c47979b613d9233838, type: 2}
         Color: {r: 1, g: 1, b: 1, a: 1}
         BuildLoad: 2
         UnTraversable: 0
