+++ Assets/Features/Building/Scripts/Datatypes/Atlas.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 80b0819e314343908c27abe48434a49b
+timeCreated: 1718708914
