+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileData/BaseTile.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4429655b7ad34b12af9a68c7a68ec558
+timeCreated: 1718709004
