+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileData/NormalTile.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Features.Building.Scripts.Datatypes.TileData
+{
+    /// <summary>
+    /// This empty class is used for the custom editor to better check the data 
+    /// </summary>
+    [Serializable]
+    public class NormalTile : BaseTile
+    {
+    }
+}
