+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileData/NormalTile.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f9abe01d3c8d405bad423d6bdeafe06e
+timeCreated: 1718709035
