+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileData/UtilityTile.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Features.Building.Scripts.Datatypes.TileData
+{
+    [Serializable]
+    public class UtilityTile : BaseTile
+    {
+        public UtilityType UtilityType;
+        public int Workload;
+    }
+}
