+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileEntry.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Features.Building.Scripts.Datatypes.TileData;
+using UnityEngine;
+
+namespace Features.Building.Scripts.Datatypes
+{
+    [Serializable]
+    public class TileEntry
+    {
+        public TileType TileType;
+        [SerializeReference]
+        public BaseTile TileData;
+    }
+}
