+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileEntry.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0a12b126133640b0a7b459cf49e64625
+timeCreated: 1718708942
