+++ Assets/Features/Building/Scripts/Datatypes/Atlas/TileType.cs
@@ -1,0 +1,8 @@
+﻿namespace Features.Building.Scripts.Datatypes
+{
+    public enum TileType
+    {
+        Normal,
+        Utility
+    }
+}
