--- Assets/Features/Building/Scripts/Datatypes/TileData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: dd1cb67aee2d4c24969843d5f3a29eff
-timeCreated: 1714987507
