+++ Assets/Features/Building/Scripts/Datatypes/Utility/UtilityType.cs
