+++ Assets/Scenes/BuildingSystem.unity
@@ -293,7 +293,7 @@ MonoBehaviour:
   <TaskManager>k__BackingField: {fileID: 1179201915}
   <BuildingManager>k__BackingField: {fileID: 940424793}
   <GridManager>k__BackingField: {fileID: 494020287}
-  <QueueManager>k__BackingField: {fileID: 9991256}
+  <QueueManager>k__BackingField: {fileID: 0}
   <StaffManager>k__BackingField: {fileID: 0}
   <FinanceManager>k__BackingField: {fileID: 1390492149}
   <GameTimeManager>k__BackingField: {fileID: 1446644599}
@@ -4231,10 +4231,6 @@ MonoBehaviour:
   <CurrentBuildableObject>k__BackingField: {fileID: 0}
   <BuildingStatuses>k__BackingField:
   - <BuildableObject>k__BackingField: {fileID: 11400000, guid: add816ad7c0618e45bc6275e18b3444f,
-      type: 2}
-    <IsUnlocked>k__BackingField: 1
-    <Category>k__BackingField: 0
-  - <BuildableObject>k__BackingField: {fileID: 11400000, guid: 331712fde85145d40b102c937e492771,
       type: 2}
     <IsUnlocked>k__BackingField: 1
     <Category>k__BackingField: 0
